+++ Source/Actors/weapons/Projectile.h
@@ -43,5 +43,5 @@ class Projectile : public Actor {
   class AABBColliderComponent* mColliderComponent;
   bool mIsDead;
   float mLifeTime;
-  uint32_t mDamage;
+  float mDamage;
 };
